[refactor] Move notification processes classes to another package
