+++ app/src/main/java/com/makentoshe/booruchan/screen/samples/model/NotificationProcess.kt
@@ -1,15 +1,11 @@
-package com.makentoshe.booruchan.notification
+package com.makentoshe.booruchan.screen.samples.model
 
 import android.annotation.SuppressLint
-import android.app.NotificationChannel
-import android.app.NotificationManager
 import android.content.Context
-import android.os.Build
 import androidx.core.app.NotificationCompat
 import com.makentoshe.booruchan.R
 import com.makentoshe.booruchan.api.Post
-import org.jetbrains.anko.doFromSdk
-import org.jetbrains.anko.notificationManager
+import com.makentoshe.booruchan.notification.NotificationBuilder
 
 class NotificationProcess(private val post: Post) {
